requirements.txt
python-docx==1.1.2
filetype==1.2.0
